services/backend/requirements.txt
musicbrainzngs==0.7.1
Django==5.0.6
django-cors-headers==4.3.1 
djangorestframework==3.15.2
mutagen==1.47.0
psycopg[binary]==3.1.19
pyinstaller==6.8.0
unidecode==1.3.8
celery==5.4.0
redis==5.0.7